requirements.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
glfw==2.9.0
kiwisolver==1.4.8
multipledispatch==1.0.0
noise==1.2.2
numpy==2.3.1
packaging==25.0
pillow==11.2.1
PyOpenGL==3.1.9
PyOpenGL-accelerate==3.1.9
pyparsing==3.2.3
pyrr==0.10.3
python-dateutil==2.9.0.post0
six==1.17.0
imgui
